% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPlumberScriptEnvVar.R
\name{copyPlumberScriptToLocation}
\alias{copyPlumberScriptToLocation}
\title{Wrapper function for copying plumber script to specified location}
\usage{
copyPlumberScriptToLocation(outputPath)
}
\description{
Wrapper function for copying plumber script to specified location
}
