% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPlumberScriptEnvVar.R
\name{createPlumberScriptEnvVar}
\alias{createPlumberScriptEnvVar}
\title{Wrapper function for starting plumber service from within Docker}
\usage{
createPlumberScriptEnvVar(filePath)
}
\description{
Wrapper function for starting plumber service from within Docker
}
