% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResourcesPage_logData.R
\name{.log_data}
\alias{.log_data}
\title{Prepare logs data for serving}
\usage{
.log_data(from, to)
}
\arguments{
\item{from}{start date for filtering, YYYY-MM-DD}

\item{to}{end date for filtering, YYYY-MM-DD}
}
\description{
Prepare logs data for serving
}
