% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadLocalFile}
\alias{loadLocalFile}
\title{Retrieve artifacts on RServe machine mapped to the docker container}
\usage{
loadLocalFile(fileName)
}
\arguments{
\item{fileName}{file name sub-string to use for search}
}
\description{
Retrieve artifacts on RServe machine mapped to the docker container
}
